+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 🗑️ **Deprecations**
 
+- Method `BigInteger::gcdMultiple()` is deprecated, use `gcdAll()` instead
 - Method `BigDecimal::getIntegralPart()` is deprecated (will be removed in 0.15, and re-introduced as returning `BigInteger` in 0.16)
 - Method `BigDecimal::getFractionalPart()` is deprecated (will be removed in 0.15, and re-introduced as returning `BigDecimal` with a different meaning in 0.16)
 - Method `BigRational::nd()` is deprecated, use `ofFraction()` instead
@@ -16,6 +17,7 @@ All notable changes to this project will be documented in this file.
 
 ✨ **New features**
 
+- New method: `BigInteger::gcdAll()` (replaces `gcdMultiple()`)
 - New method: `BigRational::clamp()`
 - New method: `BigRational::ofFraction()` (replaces `nd()`)
 - New method: `BigRational::getIntegralPart()` (replaces `quotient()`)